use int32_t instead of longs

+++ src/gui_draw.c
@@ -281,11 +281,11 @@ void draw_round_slab64k(long pos_x, long pos_y, int units_per_px, long width, lo
 {
     unsigned short drwflags_mem = RendererGetDrawFlags();
     RendererClearDrawFlags(Lb_SPRITE_OUTLINE);
-    long fill_inset = scale_ui_value_lofi(4);
+    int32_t fill_inset = scale_ui_value_lofi(4);
     /* Keep the fill out of the rounded outer edge, but cover the transparent
        interior of the larger corner and button sprites. */
-    long corner_width = scale_ui_value_lofi(12);
-    long corner_height = scale_ui_value_lofi(12);
+    int32_t corner_width = scale_ui_value_lofi(12);
+    int32_t corner_height = scale_ui_value_lofi(12);
     RendererClearDrawFlags(Lb_SPRITE_OUTLINE);
     if (style_type == ROUNDSLAB64K_LIGHT) {
         RendererAddDrawFlags(Lb_SPRITE_TRANSPAR4);
